Rank manual/mobile food search results the same way the chatbot already does
rankProviderMatches() already demoted branded items ("EGG (SNICKERS)") behind
plain whole foods, but it only ran inside foodProviderLookupService.ts, which
backs the chatbot and food-photo-import flows. The manual web search UI
(routes/v2/foodRoutes.ts) and the mobile app's per-provider search both call
searchProviderFoods() in externalFoodSearchService.ts directly, which returned
each provider's raw, unranked order -- confirmed live: searching "chicken
breast" through OpenFoodFacts came back 10/10 branded/processed results with
no plain generic entry anywhere in the list.

Move the ranking function and its ProviderFoodItem/ProviderFoodVariant types
out of foodProviderLookupService.ts into a new utils/foodRanking.ts, since
foodProviderLookupService.ts already imports searchProviderFoods FROM
externalFoodSearchService.ts and the reverse import would have made the two
modules circular. Both services now import rankProviderMatches from the new
shared module, and searchProviderFoods() ranks its combined result list once
before returning, so the manual and mobile search paths get the same
whole-food-first ordering the chatbot always had.
